@@ -11,26 +11,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.2.0",
-    "@angular/cdk": "^17.3.1",
-    "@angular/common": "^17.2.0",
-    "@angular/compiler": "^17.2.0",
-    "@angular/core": "^17.2.0",
-    "@angular/forms": "^17.2.0",
-    "@angular/material": "^17.3.1",
-    "@angular/platform-browser": "^17.2.0",
-    "@angular/platform-browser-dynamic": "^17.2.0",
-    "@angular/platform-server": "^17.2.0",
-    "@angular/router": "^17.2.0",
+    "@angular/animations": "^17.3.4",
+    "@angular/cdk": "^17.2.3",
+    "@angular/common": "^17.3.4",
+    "@angular/compiler": "^17.2.3",
+    "@angular/core": "^17.3.4",
+    "@angular/forms": "^17.2.3",
+    "@angular/material": "^17.2.3",
+    "@angular/platform-browser": "^17.3.4",
+    "@angular/platform-browser-dynamic": "^17.2.3",
+    "@angular/platform-server": "^17.2.3",
+    "@angular/router": "^17.2.3",
     "@angular/ssr": "^17.2.3",
-    "@kolkov/ngx-gallery": "^2.0.1",
     "bootstrap": "^5.3.3",
     "bootswatch": "^5.3.3",
     "express": "^4.18.2",
     "font-awesome": "^4.7.0",
     "local-storage": "^2.0.0",
     "localstorage-polyfill": "^1.0.1",
+    "ng2-file-upload": "^5.0.0",
     "ngx-bootstrap": "^12.0.0",
+    "ngx-spinner": "^16.0.2",
     "ngx-toastr": "^18.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
@@ -39,7 +40,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.2.3",
     "@angular/cli": "^17.2.3",
-    "@angular/compiler-cli": "^17.2.0",
+    "@angular/compiler-cli": "^17.2.3",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
