@@ -1,9 +1,9 @@
 export interface User{
+  roles: never[];
+knowAs: string;
     photoUrl: any;
     knownAs: any;
         username: string ;
         token: string;
     }
-    
-    
     
