@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef, HostListener, OnInit, ViewChild, viewChild } from '@angular/core';
 import { Member } from '../../_models/member';
 import { User } from '../../_models/user';
 import { AccountService } from '../../_services/account.service';
@@ -6,20 +6,30 @@ import { MembersService } from '../../_services/members.service';
 import { take } from 'rxjs';
 import { CommonModule } from '@angular/common';
 import { TabsModule } from 'ngx-bootstrap/tabs';
-import { NgxGalleryModule } from '@kolkov/ngx-gallery';
-import { FormsModule, NgModel } from '@angular/forms';
+
+import { FormsModule, NgForm, NgModel } from '@angular/forms';
+import { ToastrService } from 'ngx-toastr';
+import { HomeComponent } from '../../home/home.component';
+import { PhotoEditorComponent } from '../photo-editor/photo-editor.component';
 
 @Component({
   selector: 'app-member-edit',
   standalone: true,
-  imports: [CommonModule,TabsModule, NgxGalleryModule,FormsModule],
+  imports: [CommonModule,TabsModule,FormsModule,PhotoEditorComponent],
   templateUrl: './member-edit.component.html',
   styleUrl: './member-edit.component.css'
 })
 export class MemberEditComponent implements OnInit {
+  @ViewChild('editForm') editForm!: NgForm | undefined;
   member : Member | undefined
   user : User |undefined
-  constructor( private accountService: AccountService, private memberService :MembersService) {
+  @HostListener('window:beforeunload',['$event']) unloadNotification($event: any){
+    if(this.editForm?.dirty)
+      {
+        $event.returnValue= true
+      }
+  }
+  constructor( private accountService: AccountService, private memberService :MembersService, private toastr: ToastrService) {
     this.accountService.currentUser$.pipe(take(1)).subscribe(user => {
       if (user) {
         this.user = user
@@ -37,8 +47,14 @@ export class MemberEditComponent implements OnInit {
       });
     }
   }
-  
-    
+  updateMember(){
+    console.log(this.member)
+    this.toastr.success('Profile updated successfully')
+    this.editForm?.reset(this.member)
+  }
+
+
+   
 }
   
   
