@@ -3,10 +3,10 @@ import { Member } from '../../_models/member';
 import { MembersService } from '../../_services/members.service';
 import { CommonModule, NgFor } from '@angular/common';
 import { MemberCardComponent } from "../member-card/member-card.component";
-import { Router } from 'express';
 import { RouterModule, RouterOutlet } from '@angular/router';
 import { BrowserAnimationsModule, NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { HomeComponent } from '../../home/home.component';
+import { Observable, map } from 'rxjs';
 
 @Component({
     selector: 'app-member-list',
@@ -17,18 +17,12 @@ import { HomeComponent } from '../../home/home.component';
 })
 export class MemberListComponent implements OnInit {
 
- members: Member[] | undefined
+  members$: Observable<Member[]> | undefined 
   constructor(private memberService: MembersService) {
   
   }
   ngOnInit(): void {
-      this.loadMembers()
+      this.members$= this.memberService.getMembers()
   }
-  loadMembers(){
-    this.memberService.getMembers().subscribe(members=>
-      {
-        this.members = members
-      })
-  }
-
+  
 }
