@@ -3,12 +3,15 @@ using DatingApp.Data;
 using DatingApp.DTOs;
 using DatingApp.Interfaces;
 using DatingApp.Models;
+using DatingApp.Services;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace DatingApp.Controllers
@@ -19,27 +22,118 @@ namespace DatingApp.Controllers
     {
         private readonly IUserRepository _userRepository;
         private readonly IMapper _mapper;
-        public UsersController(IUserRepository userRepository, IMapper mapper) 
-        { 
-          _userRepository=userRepository;
-            _mapper=mapper;
+        private readonly IPhotoService _photoService;
+
+        public UsersController(IUserRepository userRepository, IMapper mapper, IPhotoService photoService)
+        {
+            _userRepository = userRepository;
+            _mapper = mapper;
+            _photoService = photoService;
         }
 
         [HttpGet]
-        
-        public async Task<ActionResult<IEnumerable<MemberDto>>> GetUsers() 
+
+        public async Task<ActionResult<IEnumerable<MemberDto>>> GetUsers()
         {
             var users = await _userRepository.GetMembersAsync();
 
             return Ok(users);
         }
 
-        
-        [HttpGet("{username}")]
+
+        [HttpGet("{username}", Name = "GetUser")]
         public async Task<ActionResult<MemberDto>> GetUser(string username)
         {
             return await _userRepository.GetMemberAsync(username);
-            
+
         }
+
+        [HttpPut]
+        public async Task<ActionResult> UpdateUser(MemberUpdateDto memberUpdateDto)
+        {
+
+            var user = await _userRepository.GetUserByUsernameAsync(username: User.GetUserName());
+
+            _mapper.Map(memberUpdateDto, user);
+            _userRepository.Upadate(user);
+
+            if (await _userRepository.SaveAllAsync()) return NoContent();
+
+            return BadRequest("Failed to update user");
+        }
+
+        [HttpPost("add-photo")]
+        public async Task<ActionResult<PhotoDto>> AddPhoto(IFormFile File)
+        {
+            var user = await _userRepository.GetUserByUsernameAsync(username: User.GetUserName());
+
+            var result = await _photoService.AddPhotoAsync(File);
+
+            if (result.Error != null) return BadRequest(result.Error.Message);
+
+            var photo = new Photo
+            {
+                Url = result.SecureUrl.AbsoluteUri,
+                PublicId = result.PublicId,
+            };
+            if (user == null)
+            {
+                // Xử lý trường hợp không tìm thấy người dùng
+                return BadRequest("User not found");
+            }
+            if (user.Photos.Count == 0)
+            {
+                photo.IsMain = true;
+            }
+            user.Photos.Add(photo);
+
+            if (await _userRepository.SaveAllAsync())
+            {
+                //return _mapper.Map<PhotoDto>(photo);
+                return CreatedAtRoute("GetUser", new { username = user.Username }, _mapper.Map<PhotoDto>(photo));
+            }
+            return BadRequest("Problem adding photo");
+        }
+
+        [HttpPut("set-main-photo/{photoId}")]
+        public async Task<ActionResult> SetMainPhoto(int photoId)
+        {
+            var user = await _userRepository.GetUserByUsernameAsync(User.GetUserName());
+            var photo = user.Photos.FirstOrDefault(x=>x.Id == photoId);
+            if (photo == null) return NotFound();
+            if (photo.IsMain) return BadRequest("This is already your main photo");
+
+            var currentMain = user.Photos.FirstOrDefault(x=>x.IsMain);
+            if (currentMain != null)
+            {
+                currentMain.IsMain = false;
+            }
+            photo.IsMain = true;
+            if(await _userRepository.SaveAllAsync()) return NoContent();
+
+            return BadRequest("Problems setting main photo");
+
+        }
+
+        [HttpDelete("delete-photo/{photoId}")]
+        public async Task<ActionResult> DeletePhoto(int photoId)
+        {
+            var user = await _userRepository.GetUserByUsernameAsync(User.GetUserName());
+            var photo = user.Photos.FirstOrDefault(x=>x.Id == photoId);
+
+            if(photo == null) return NotFound();
+            if (photo.IsMain) return BadRequest("you cannot delete your main photo");
+            if(photo.PublicId !=null)
+            {
+                var result = await _photoService.DeletePhotoAsync(photo.PublicId);
+                if (result.Error != null) return BadRequest(result.Error.Message);
+               
+            }
+            user.Photos.Remove(photo);
+            if (await _userRepository.SaveAllAsync()) return Ok();
+            return BadRequest("Faild to delete photo");
+
+        }
+        
     }
 }
